feat: move utils away from root directory
